fix: rename functions, replace mock const in mock holder
