+++ src/view/point-list-view.js
@@ -1,11 +1,11 @@
 import AbstractView from '../framework/view/abstract-view';
 
-function createPointListViewTemplate() {
+function createPointListTemplate() {
   return '<ul class="trip-events__list"></ul>';
 }
 
 export default class PointListView extends AbstractView {
   get template() {
-    return createPointListViewTemplate();
+    return createPointListTemplate();
   }
 }
