+++ src/view/trip-info-view.js
@@ -1,6 +1,6 @@
 import AbstractView from '../framework/view/abstract-view';
 
-function createTripInfoViewTemplate() {
+function createTripInfoTemplate() {
   return `<section class="trip-main__trip-info  trip-info">
   <div class="trip-info__main">
     <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
@@ -16,6 +16,6 @@ function createTripInfoViewTemplate() {
 
 export default class TripInfoView extends AbstractView {
   get template() {
-    return createTripInfoViewTemplate();
+    return createTripInfoTemplate();
   }
 }
